Move the model evaluator package + update models for rewards
