+++ bundles/org.palladiosimulator.analyzer.slingshot.behavior.spd/src/org/palladiosimulator/analyzer/slingshot/behavior/spd/interpreter/entity/model/ImprovedQLearningModelEvaluator.java
@@ -1,16 +1,17 @@
-package org.palladiosimulator.analyzer.slingshot.behavior.spd.interpreter.models;
+package org.palladiosimulator.analyzer.slingshot.behavior.spd.interpreter.entity.model;
 
 import java.util.ArrayList;
 import java.util.List;
 
 import org.palladiosimulator.analyzer.slingshot.behavior.spd.interpreter.entity.aggregator.ModelAggregatorWrapper;
+import org.palladiosimulator.analyzer.slingshot.behavior.spd.interpreter.entity.model.reward.RewardEvaluator;
 import org.palladiosimulator.spd.adjustments.models.ImprovedQLearningModel;
 
 public class ImprovedQLearningModelEvaluator extends LearningBasedModelEvaluator {
 
     public ImprovedQLearningModelEvaluator(ImprovedQLearningModel model,
-            List<ModelAggregatorWrapper<?>> stimuliListeners) {
-        super(stimuliListeners);
+            List<ModelAggregatorWrapper<?>> stimuliListeners, RewardEvaluator rewardEvaluator) {
+        super(stimuliListeners, rewardEvaluator);
     }
 
     @Override
