+++ bundles/org.palladiosimulator.analyzer.slingshot.behavior.spd/src/org/palladiosimulator/analyzer/slingshot/behavior/spd/interpreter/entity/model/LearningBasedModelEvaluator.java
@@ -1,15 +1,17 @@
-package org.palladiosimulator.analyzer.slingshot.behavior.spd.interpreter.models;
+package org.palladiosimulator.analyzer.slingshot.behavior.spd.interpreter.entity.model;
 
 import java.util.List;
 
 import org.palladiosimulator.analyzer.slingshot.behavior.spd.interpreter.entity.aggregator.ModelAggregatorWrapper;
+import org.palladiosimulator.analyzer.slingshot.behavior.spd.interpreter.entity.model.reward.RewardEvaluator;
 import org.palladiosimulator.analyzer.slingshot.monitor.data.events.MeasurementMade;
 
 public abstract class LearningBasedModelEvaluator extends ModelEvaluator {
 
     protected final List<ModelAggregatorWrapper<?>> aggregatorList;
 
-    public LearningBasedModelEvaluator(List<ModelAggregatorWrapper<?>> stimuliListeners) {
+    public LearningBasedModelEvaluator(List<ModelAggregatorWrapper<?>> stimuliListeners,
+            RewardEvaluator rewardEvaluator) {
         super();
         this.aggregatorList = stimuliListeners;
     }
